[#show_all_generated_docs] add new option (replace old documents by the latest?)

+++ components/com_emundus/models/evaluation.php
@@ -2403,6 +2403,11 @@ class EmundusModelEvaluation extends JModelList {
 
     /// generate letters
     public function generateLetters($fnums, $templates, $canSee, $showMode, $mergeMode) {
+        $eMConfig = JComponentHelper::getParams('com_emundus');
+
+        /* replace old documents by the latest */
+        $replace_document = $eMConfig->get('export_replace_doc', 0);
+
         $tmp_path = JPATH_BASE . DS . 'tmp' . DS;
 
         require_once(JPATH_BASE.DS.'components'.DS.'com_emundus' . DS . 'models' . DS . 'evaluation.php');
@@ -2456,13 +2461,15 @@ class EmundusModelEvaluation extends JModelList {
                 // get attachment info
                 $attachInfo = $_mFile->getAttachmentInfos($letter->attachment_id);
 
-                /// before to generate letter, refresh all previous generated letters of current day
-                $refreshQuery = 'DELETE FROM #__emundus_uploads WHERE #__emundus_uploads.attachment_id = ' . $attachInfo['id'] .
-                                    ' AND DATE(#__emundus_uploads.timedate) = current_date() ' .
-                                        ' AND #__emundus_uploads.fnum = ' . $fnum;
+                /* before to generate letter, refresh all previous generated letters of current day - if $replace_document = true */
+                if($replace_document == 1) {
+                    $refreshQuery = 'DELETE FROM #__emundus_uploads WHERE #__emundus_uploads.attachment_id = ' . $attachInfo['id'] .
+                        ' AND DATE(#__emundus_uploads.timedate) = current_date() ' .
+                        ' AND #__emundus_uploads.fnum = ' . $fnum;
 
-                $this->_db->setQuery($refreshQuery);
-                $this->_db->execute();
+                    $this->_db->setQuery($refreshQuery);
+                    $this->_db->execute();
+                }
 
                 $type = $letter->template_type;
 
@@ -2542,12 +2549,12 @@ class EmundusModelEvaluation extends JModelList {
 
                                 /// reupdate in database
                                 $upId = $_mFile->addAttachment($fnum, $name, $fnumInfo[$fnum]['applicant_id'], $fnumInfo[$fnum]['campaign_id'], $letter->attachment_id, $attachInfo['description'], $canSee);
-                                $res->files[] = array('filename' => $name, 'upload' => $upId, 'url' => $original_url);
+                                $res->files[] = array('filename' => $name, 'upload' => $upId, 'url' => $original_url, 'type' => $attachInfo['id']);
                             } else {
                                 if (copy($file, $path_name) and copy($file, $original_name)) {
                                     $upId = $_mFile->addAttachment($fnum, $name, $fnumInfo[$fnum]['applicant_id'], $fnumInfo[$fnum]['campaign_id'], $letter->attachment_id, $attachInfo['description'], $canSee);
 
-                                    $res->files[] = array('filename' => $name, 'upload' => $upId, 'url' => $original_url);
+                                    $res->files[] = array('filename' => $name, 'upload' => $upId, 'url' => $original_url, 'type' => $attachInfo['id']);
                                 }
                             }
                         } else {
@@ -2682,7 +2689,7 @@ class EmundusModelEvaluation extends JModelList {
 
                             $pdf->Output($path_name, 'F');
                             $pdf->Output($original_name, 'F');
-                            $res->files[] = array('filename' => $name, 'upload' => $upId, 'url' => $original_url);
+                            $res->files[] = array('filename' => $name, 'upload' => $upId, 'url' => $original_url, 'type' => $attachInfo['id']);
                         }
                         unset($pdf, $path_name, $name, $url, $upIdn);
                         unset($pdf, $original_name, $name, $original_url, $upIdn);
@@ -2917,7 +2924,7 @@ class EmundusModelEvaluation extends JModelList {
                                 }
 
                                 $upId = $_mFile->addAttachment($fnum, $filename, $fnumInfo[$fnum]['applicant_id'], $fnumInfo[$fnum]['campaign_id'], $letter->attachment_id, $attachInfo['description'], $canSee);
-                                $res->files[] = array('filename' => $filename, 'upload' => $upId, 'url' => $original_url);
+                                $res->files[] = array('filename' => $filename, 'upload' => $upId, 'url' => $original_url, 'type' => $attachInfo['id']);
                             }
                         } catch (Exception $e) {
                             $res->status = false;
@@ -3095,7 +3102,7 @@ class EmundusModelEvaluation extends JModelList {
                                 copy($original_name, $path_name);
 
                                 $upId = $_mFile->addAttachment($fnum, $filename, $fnumInfo[$fnum]['applicant_id'], $fnumInfo[$fnum]['campaign_id'], $letter->attachment_id, $attachInfo['description'], $canSee);
-                                $res->files[] = array('filename' => $filename, 'upload' => $upId, 'url' => $original_url);
+                                $res->files[] = array('filename' => $filename, 'upload' => $upId, 'url' => $original_url, 'type' => $attachInfo['id']);
 
                             } else {
                                 $upId = $_mFile->addAttachment($fnum, $filename, $fnumInfo[$fnum]['applicant_id'], $fnumInfo[$fnum]['campaign_id'], $letter->attachment_id, $attachInfo['description'], $canSee);
@@ -3106,7 +3113,7 @@ class EmundusModelEvaluation extends JModelList {
 
                                 copy($original_name, $path_name);
 
-                                $res->files[] = array('filename' => $filename, 'upload' => $upId, 'url' => $original_url);
+                                $res->files[] = array('filename' => $filename, 'upload' => $upId, 'url' => $original_url, 'type' => $attachInfo['id']);
                             }
                             break;
                         }
@@ -3130,9 +3137,11 @@ class EmundusModelEvaluation extends JModelList {
             $duplicateAttachments = $this->_db->loadObjectList();
 
             /// remove unnecessary records for same attachment id in database
-            $deleteDuplicateAttachmentsQuery = 'DELETE FROM #__emundus_uploads WHERE #__emundus_uploads.id NOT IN ( ' . implode(',', $_ids) . ' ) AND #__emundus_uploads.fnum = ' . $fnum;
-            $this->_db->setQuery($deleteDuplicateAttachmentsQuery);
-            $this->_db->execute();
+            if($replace_document == 1) {
+                $deleteDuplicateAttachmentsQuery = 'DELETE FROM #__emundus_uploads WHERE #__emundus_uploads.id NOT IN ( ' . implode(',', $_ids) . ' ) AND #__emundus_uploads.fnum = ' . $fnum;
+                $this->_db->setQuery($deleteDuplicateAttachmentsQuery);
+                $this->_db->execute();
+            }
 
 //            foreach($duplicateAttachments as $attachment) {
 //                unlink(EMUNDUS_PATH_ABS . $attachment->user_id . DS . $attachment->filename);
@@ -3141,28 +3150,56 @@ class EmundusModelEvaluation extends JModelList {
         }
 
         /// lastly, we get all records in jos_emundus_uploads
-        unset($res->files);
+        if($replace_document == 1) {
+            unset($res->files);
 
-        $availableFilesName = [];
+            $availableFilesName = [];
 
-        $getAllUploadsQuery = 'SELECT #__emundus_uploads.* FROM #__emundus_uploads WHERE #__emundus_uploads.fnum IN (' . implode(',', $fnum_Array) . ') AND DATE(#__emundus_uploads.timedate) = current_date() AND #__emundus_uploads.attachment_id IN (' . implode(',', $templates) . ' )';
-        $this->_db->setQuery($getAllUploadsQuery);
-        $_upAttachments = $this->_db->loadObjectList();
+            $getAllUploadsQuery = 'SELECT #__emundus_uploads.* FROM #__emundus_uploads WHERE #__emundus_uploads.fnum IN (' . implode(',', $fnum_Array) . ') AND DATE(#__emundus_uploads.timedate) = current_date() AND #__emundus_uploads.attachment_id IN (' . implode(',', $templates) . ' )';
+            $this->_db->setQuery($getAllUploadsQuery);
+            $_upAttachments = $this->_db->loadObjectList();
 
-        foreach($_upAttachments as $_upload) {
-            $res->files[] = array('filename' => $_upload->filename, 'upload' => $_upload->id, 'url' => JURI::base() . EMUNDUS_PATH_REL . $_upload->user_id . DS);
-            $availableFilesName[] = $_upload->filename;
-        }
+            foreach ($_upAttachments as $_upload) {
+                $res->files[] = array('filename' => $_upload->filename, 'upload' => $_upload->id, 'url' => JURI::base() . EMUNDUS_PATH_REL . $_upload->user_id . DS);
+                $availableFilesName[] = $_upload->filename;
+            }
 
-        foreach($fnum_Array as $fnum) {
-            $fnumInfo = $_mFile->getFnumsTagsInfos([$fnum]);
-            $files = glob(EMUNDUS_PATH_ABS . 'tmp' . DS . $fnumInfo[$fnum]['applicant_name'] .'_' . $fnumInfo[$fnum]['applicant_id'] . DS . '*');
+            foreach ($fnum_Array as $fnum) {
+                $fnumInfo = $_mFile->getFnumsTagsInfos([$fnum]);
+                $files = glob(EMUNDUS_PATH_ABS . 'tmp' . DS . $fnumInfo[$fnum]['applicant_name'] . '_' . $fnumInfo[$fnum]['applicant_id'] . DS . '*');
 
-            foreach($files as $_f) {
-                $fname = end(explode('/', $_f));
-                if(!in_array($fname, $availableFilesName)) {
-                    unlink($_f);
+                foreach ($files as $_f) {
+                    $fname = end(explode('/', $_f));
+                    if (!in_array($fname, $availableFilesName)) {
+                        unlink($_f);
+                    }
                 }
+            }
+        } else {
+            $_idList = array();
+            /* get max upload id for each fnum and for each attachment id and in currrent_date() */
+            foreach($templates as $tmpl) {
+                $getLastUploadIdQuery = 'SELECT MAX(#__emundus_uploads.id) as uid
+                                            FROM #__emundus_uploads 
+                                                WHERE #__emundus_uploads.fnum IN (' . implode(',', $fnum_Array) . ') 
+                                                    AND DATE(#__emundus_uploads.timedate) = current_date() 
+                                                        AND #__emundus_uploads.attachment_id IN (' . $tmpl . ' ) 
+                                                            GROUP BY fnum';
+                $this->_db->setQuery($getLastUploadIdQuery);
+                $_idList[$tmpl] = $this->_db->loadAssocList();
+            }
+
+            $raw = array_reduce($_idList, 'array_merge', array());
+            $out = array();
+
+            foreach($raw as $r) { $out[] = $r['uid']; }
+            
+            /* from $out, get the filename of each */
+            unset($res->files);
+            $getFiles = $_mFile->getAttachmentsById($out);
+
+            foreach($getFiles as $f) {
+                $res->files[] = array('filename' => $f['filename'], 'upload' => $f['id'], 'url' => JURI::base() . EMUNDUS_PATH_REL . $f['user_id'] . DS, 'type' => $f['attachment_id']);
             }
         }
 
@@ -3417,23 +3454,34 @@ class EmundusModelEvaluation extends JModelList {
         }
 
         // build the recap table (just get generated documents of current date)
-        $query = 'SELECT #__emundus_uploads.attachment_id, COUNT(#__emundus_uploads.attachment_id) AS _count 
+        if($replace_document == 1) {
+            $query = 'SELECT #__emundus_uploads.attachment_id, COUNT(#__emundus_uploads.attachment_id) AS _count 
                         FROM #__emundus_uploads
-                            WHERE #__emundus_uploads.fnum in (' . implode(',' , $available_fnums) . ') 
+                            WHERE #__emundus_uploads.fnum in (' . implode(',', $available_fnums) . ') 
                                 AND #__emundus_uploads.attachment_id IN (' . implode(',', $templates) . ')
                                     AND DATE(#__emundus_uploads.timedate) = current_date()
                                         GROUP BY #__emundus_uploads.attachment_id';
 
-        $this->_db->setQuery($query);
-
-        $document_count = $this->_db->loadAssocList();
-
-        $res->recapitulatif_count = [];
-
-        foreach($document_count as $key => $document) {
-            $query = "SELECT #__emundus_setup_attachments.value FROM #__emundus_setup_attachments WHERE #__emundus_setup_attachments.id = " . $document['attachment_id'];
             $this->_db->setQuery($query);
-            $res->recapitulatif_count[] = array('document' => $this->_db->loadResult(), 'count' => $document['_count']);
+            $document_count = $this->_db->loadAssocList();
+            $res->recapitulatif_count = [];
+
+            foreach ($document_count as $key => $document) {
+                $query = "SELECT #__emundus_setup_attachments.value FROM #__emundus_setup_attachments WHERE #__emundus_setup_attachments.id = " . $document['attachment_id'];
+                $this->_db->setQuery($query);
+                $res->recapitulatif_count[] = array('document' => $this->_db->loadResult(), 'count' => $document['_count']);
+            }
+        } else {
+            $res->recapitulatif_count = [];
+            $raw = [];
+
+            foreach($res->files as $file) { $raw[$file['type']] += 1; }
+            
+           foreach($raw as $k => $v) {
+               $query = "SELECT #__emundus_setup_attachments.value FROM #__emundus_setup_attachments WHERE #__emundus_setup_attachments.id = " . $k;
+               $this->_db->setQuery($query);
+               $res->recapitulatif_count[] = array('document' => $this->_db->loadResult(), 'count' => $v);
+           }
         }
 
         return json_encode($res);
